client/requirements.txt
Flask>=3.0,<4.0
paho-mqtt>=1.6.1,<2.0
